+++ src/Billing/BillingRouter.js
@@ -2,8 +2,8 @@
 const stripe = require('stripe')(process.env.STRIPE_SECRET);
 
 // Get tenant with property Id
-const getCustomer = (req, res) => {
+const getCustomerCC = (req, res) => {
   const { id } = req.params;
 };
 
-module.exports = { getCustomer };
+module.exports = { getCustomerCC };
