+++ src/Billing/index.js
@@ -3,8 +3,8 @@ const express = require('express');
 const router = express.Router();
 
 // import the route modules
-const { getCustomer } = require('./BillingRouter');
+const { getCustomerCC } = require('./BillingRouter');
 
-router.get('/get/:id', getCustomer); // Get CC and Rent history
+router.get('/get/:id', getCustomerCC); // Get CC and Rent history
 
 module.exports = router;
