+++ src/Properties/index.js
@@ -6,6 +6,7 @@ const router = express.Router();
 const {
   // propertieslsdb,
   // lsdb,
+  scanF,
   properties,
   addProperty,
   // addTenant,
@@ -14,7 +15,8 @@ const {
 } = require('./propertiesRouter');
 
 // router.get('/lsdb', propertieslsdb);
-router.get('/all', properties);
+router.get('/all', properties); // Only returns properties
+router.get('/all', scanF); // General query function that returns any item with the same admin id
 router.post('/add', addProperty);
 // router.post('/addtenant/', addTenant);
 // router.post('/lsdb', lsdb);
